+++ CHANGELOG.md
@@ -11,3 +11,4 @@
   - feat(Domain): Add basic logic of subdomain enumeration
   - feat(Asset, Domain): Capture enter key & Make all sorts of scanning async
   - feat(Port): Add basic logic of port scanning
+  - style(Asset, Domain, Port): Add tooltips for buttons